+++ test_e2e/test-dest.sh
@@ -24,6 +24,7 @@ all_tests=(
   'couchbase.sh'
   'delta-table.sh'
   'dropbox.sh'
+  'duckdb.sh'
   'elasticsearch.sh'
   'gcs.sh'
   'kafka-local.sh'