+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.1.2
+
+### Features
+
+* **Add DuckDB destination connector** Adds support storing artifacts in a local DuckDB database.
+* **Add MotherDuck destination connector** Adds support storing artifacts in MotherDuck database.
+
 ## 0.1.1-dev1
 
 ### Enhancements